+++ src/widgets/profile/ui/index.ts
@@ -1,1 +1,1 @@
-export * from "./ProfileOverView";
+export * from "./dashboard";
